Switched to LLVM 3.8 for master.
Also moved stuff out of master into custom_tests for now. Have to cd into custom_tests directory and ../.. back to the preprocessor right now to make it work because of the way it searches for files. But this will have to be addressed anyway when we have a library directory.